requirements.txt
APScheduler==3.6.3
bs4==0.0.1
flake8==3.8.3
psycopg2==2.8.5
python-telegram-bot==12.7
requests==2.22.0
SQLAlchemy==1.3.17
peewee==3.13.3
flake8-docstrings==1.5.0
black==19.10b0
pytest==5.4.3
pytest-cov==2.10.1
